new density-based balancer
This balancer has the nice property that it works when data is duplicated. It still doesn't seem to generalize well, though.

One thing to note about this balancer: it would ideally work with temperature=1, but the VAE puts log density deltas of ~300 nats between a sample and its nearest neighbor. I am going to try re-training a VAE that is forced to put similar samples in an overlapping part of the prior.